backend/requirements.txt
# Основной веб-фреймворк
fastapi>=0.100.0,<0.101.0

# ASGI сервер для запуска FastAPI приложений
uvicorn[standard]>=0.23.0,<0.24.0

# Работа с базой данных
sqlalchemy>=2.0.0,<3.0.0

# Драйвер для PostgreSQL (выберите нужный для вашей БД)
# Для PostgreSQL
asyncpg>=0.28.0,<0.29.0
# Альтернатива для PostgreSQL (может требовать psycopg2-binary)
# psycopg2-binary>=2.9.0,<3.0.0
#
# Для MySQL (если выберете MySQL)
# aiomysql>=0.2.0,<0.3.0
# pymysql>=1.1.0,<2.0.0
# Основной веб-фреймворк
fastapi>=0.100.0,<0.101.0

# ASGI сервер для запуска FastAPI приложений
uvicorn[standard]>=0.23.0,<0.24.0

# Работа с базой данных
sqlalchemy>=2.0.0,<3.0.0

# Асинхронный драйвер для SQLite
aiosqlite>=0.19.0,<0.20.0

# Валидация данных и сериализация
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0 # Для работы с настройками/переменными окружения

# Библиотека для Telegram Bot API
python-telegram-bot>=20.4,<21.0

# Для хэширования и проверки подписи initData
# (hashlib и hmac входят в стандартную библиотеку)
# Валидация данных и сериализация
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0 # Для работы с настройками/переменными окружения

# Библиотека для Telegram Bot API
python-telegram-bot>=20.4,<21.0

# Для хэширования и проверки подписи initData
# (hashlib и hmac входят в стандартную библиотеку, но могут понадобиться дополнительные утилиты)